requirements.txt
python-dotenv==1.0.1
pytest~=8.2.2
selenium~=4.22.0